+++ packages/anthropic/src/anthropic-messages-options.ts
@@ -2,21 +2,23 @@ import { z } from 'zod/v4';
 
 // https://docs.claude.com/en/docs/about-claude/models/overview
 export type AnthropicMessagesModelId =
-  | 'claude-haiku-4-5'
-  | 'claude-haiku-4-5-20251001'
-  | 'claude-sonnet-4-5'
-  | 'claude-sonnet-4-5-20250929'
-  | 'claude-opus-4-1'
-  | 'claude-opus-4-0'
-  | 'claude-sonnet-4-0'
-  | 'claude-opus-4-1-20250805'
-  | 'claude-opus-4-20250514'
-  | 'claude-sonnet-4-20250514'
-  | 'claude-3-7-sonnet-latest'
-  | 'claude-3-7-sonnet-20250219'
-  | 'claude-3-5-haiku-latest'
   | 'claude-3-5-haiku-20241022'
+  | 'claude-3-5-haiku-latest'
+  | 'claude-3-7-sonnet-20250219'
+  | 'claude-3-7-sonnet-latest'
   | 'claude-3-haiku-20240307'
+  | 'claude-haiku-4-5-20251001'
+  | 'claude-haiku-4-5'
+  | 'claude-opus-4-0'
+  | 'claude-opus-4-1-20250805'
+  | 'claude-opus-4-1'
+  | 'claude-opus-4-20250514'
+  | 'claude-opus-4-5'
+  | 'claude-opus-4-5-20251101'
+  | 'claude-sonnet-4-0'
+  | 'claude-sonnet-4-20250514'
+  | 'claude-sonnet-4-5-20250929'
+  | 'claude-sonnet-4-5'
   | (string & {});
 
 /**
@@ -62,6 +64,15 @@ export const anthropicProviderOptions = z.object({
    * This allows you to deactivate reasoning inputs for models that do not support them.
    */
   sendReasoning: z.boolean().optional(),
+
+  /**
+   * Determines how structured outputs are generated.
+   *
+   * - `outputFormat`: Use the `output_format` parameter to specify the structured output format.
+   * - `jsonTool`: Use a special 'json' tool to specify the structured output format.
+   * - `auto`: Use 'outputFormat' when supported, otherwise use 'jsonTool' (default).
+   */
+  structuredOutputMode: z.enum(['outputFormat', 'jsonTool', 'auto']).optional(),
 
   /**
    * Configuration for enabling Claude's extended thinking.
@@ -142,6 +153,11 @@ export const anthropicProviderOptions = z.object({
    * @default true
    */
   toolStreaming: z.boolean().optional(),
+
+  /**
+   * @default 'high'
+   */
+  effort: z.enum(['low', 'medium', 'high']).optional(),
 });
 
 export type AnthropicProviderOptions = z.infer<typeof anthropicProviderOptions>;