+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/vue
 
+## 3.0.0-beta.116
+
+### Patch Changes
+
+- Updated dependencies [4ece5f9]
+  - ai@6.0.0-beta.116
+
+## 3.0.0-beta.115
+
+### Patch Changes
+
+- Updated dependencies [7da02d2]
+  - ai@6.0.0-beta.115
+
 ## 3.0.0-beta.114
 
 ### Patch Changes